backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
loguru
requests
google-genai
google-api-python-client